refactor: 필요없는 컴포넌트 제거 & 파일 위치 이동
