--- components/Skeleton/CardListSkeleton/CardListSkeleton.module.css
@@ -1,78 +1,0 @@
-.cardListSkeleton {
-  display: flex;
-  justify-content: center;
-}
-
-.container {
-  display: grid;
-  grid-template-columns: repeat(3, 34rem);
-  column-gap: 2rem;
-  row-gap: 2.5rem;
-}
-
-.cardItemSkeleton {
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  background-color: #fff;
-  width: 34rem;
-  height: 33.4rem;
-  box-shadow: 0rem 0.5rem 1rem rgba(0, 0, 0, 0.08);
-  border-radius: 1.5rem;
-}
-
-.imgContainer {
-  width: 100%;
-  height: 60%;
-  background-color: #dadada;
-  border-radius: 1.5rem;
-}
-
-.contentContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  padding: 1.5rem 2rem;
-}
-
-.timeDiff {
-  width: 30rem;
-  height: 1.8rem;
-  background-color: #dadada;
-  border-radius: 1.5rem;
-}
-
-.description {
-  width: 30rem;
-  height: 5.2rem;
-  background-color: #dadada;
-  border-radius: 1.5rem;
-}
-
-.createdAt {
-  width: 12rem;
-  height: 1.8rem;
-  background-color: #dadada;
-  border-radius: 1.5rem;
-}
-
-/* Tablet 화면 크기 (768px 이상 ~ 1199px 이하) */
-@media screen and (min-width: 768px) and (max-width: 1199px) {
-  .container {
-    grid-template-columns: repeat(2, 34rem);
-    column-gap: 2.4rem;
-  }
-}
-
-/* Mobile 화면 크기 (375px 이상 ~ 767px 이하) */
-@media screen and (min-width: 375px) and (max-width: 767px) {
-  .container {
-    grid-template-columns: 32.5rem;
-    row-gap: 2rem;
-  }
-
-  .cardItemSkeleton {
-    width: 32.5rem;
-    height: 32.7rem;
-  }
-}
