--- components/Skeleton/FolderInfoSkeleton/FolderInfoSkeleton.module.css
@@ -1,65 +1,0 @@
-.folderInfoSkeleton {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 3.6rem;
-  padding-top: 2rem;
-  padding-bottom: 6rem;
-}
-
-.profileInfo {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 1.2rem;
-}
-
-.profileImg {
-  background-color: #dadada;
-  width: 6rem;
-  height: 6rem;
-  border-radius: 50%;
-}
-
-.ownerName {
-  background-color: #dadada;
-  width: 10rem;
-  height: 2rem;
-  border-radius: 1.6rem;
-}
-
-.folderName {
-  background-color: #dadada;
-  width: 30rem;
-  height: 4rem;
-  border-radius: 1.6rem;
-}
-
-/* Tablet 화면 크기 (768px 이상 ~ 1199px 이하) */
-@media screen and (min-width: 768px) and (max-width: 1199px) {
-  .folderInfoSkeleton {
-    gap: 3.5rem;
-  }
-}
-
-/* Mobile 화면 크기 (375px 이상 ~ 767px 이하) */
-@media screen and (min-width: 375px) and (max-width: 767px) {
-  .folderInfoSkeleton {
-    gap: 4rem;
-  }
-
-  .profileImg {
-    width: 4rem;
-    height: 4rem;
-  }
-
-  .ownerName {
-    height: 1.4rem;
-  }
-
-  .folderName {
-    height: 2.79rem;
-  }
-}
